ENGR00138188 runtime: add a fake 3G modem path
Add a fake port in runtime check rild.since if return NULL to ril-reference,
it will cause rild init failed, so the framwork can't connect to rild,
and will hold a wake lock, this will cause system can't into suspend,
it will in early suspend 15 seconds.

so add a fake port to fix this.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>